Generalize the blocks as well.
