+++ MagicDoorknob/src/main/java/com/tomboshoven/minecraft/magicdoorknob/blocks/BlockMagicDoorway.java
@@ -6,25 +6,19 @@ import com.tomboshoven.minecraft.magicdoorknob.blocks.tileentities.TileEntityMag
 import com.tomboshoven.minecraft.magicdoorknob.items.ItemMagicDoorknob;
 import com.tomboshoven.minecraft.magicdoorknob.properties.PropertyTexture;
 import mcp.MethodsReturnNonnullByDefault;
-import net.minecraft.block.Block;
 import net.minecraft.block.SoundType;
-import net.minecraft.block.material.MapColor;
-import net.minecraft.block.material.MaterialTransparent;
 import net.minecraft.block.properties.PropertyBool;
 import net.minecraft.block.properties.PropertyEnum;
 import net.minecraft.block.state.BlockFaceShape;
 import net.minecraft.block.state.BlockStateContainer;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.particle.ParticleManager;
 import net.minecraft.client.renderer.BlockModelShapes;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.client.renderer.texture.TextureMap;
 import net.minecraft.entity.Entity;
 import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.BlockRenderLayer;
 import net.minecraft.util.EnumFacing;
-import net.minecraft.util.IStringSerializable;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
@@ -41,7 +35,7 @@ import java.util.List;
 
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
-public class BlockMagicDoorway extends Block {
+public class BlockMagicDoorway extends BlockMagicDoorwayPartBase {
     /**
      * Property describing which part of the doorway is being represented by this block.
      */
@@ -76,8 +70,6 @@ public class BlockMagicDoorway extends Block {
      * This is typically not necessary. Use Blocks.blockMagicDoorway instead.
      */
     BlockMagicDoorway() {
-        super(new MaterialTransparent(MapColor.AIR));
-
         // By default, the doorway is not open in any direction
         setDefaultState(
                 blockState.getBaseState()
@@ -88,50 +80,6 @@ public class BlockMagicDoorway extends Block {
     }
 
     @Override
-    public boolean addDestroyEffects(World world, BlockPos pos, ParticleManager manager) {
-        // Skip all block breaking textures
-        return true;
-    }
-
-    @Override
-    public int getLightValue(IBlockState state, IBlockAccess world, BlockPos pos) {
-        TileEntity tileEntity = world.getTileEntity(pos);
-        if (tileEntity instanceof TileEntityMagicDoorway) {
-            return ((TileEntityMagicDoorway) tileEntity).getBaseBlockState().getLightValue(world, pos);
-        }
-        return super.getLightValue(state, world, pos);
-    }
-
-    @Override
-    public int getLightOpacity(IBlockState state, IBlockAccess world, BlockPos pos) {
-        TileEntity tileEntity = world.getTileEntity(pos);
-        if (tileEntity instanceof TileEntityMagicDoorway) {
-            return ((TileEntityMagicDoorway) tileEntity).getBaseBlockState().getLightOpacity(world, pos);
-        }
-        return super.getLightOpacity(state, world, pos);
-    }
-
-    @Override
-    public float getBlockHardness(IBlockState blockState, World worldIn, BlockPos pos) {
-        TileEntity tileEntity = worldIn.getTileEntity(pos);
-        if (tileEntity instanceof TileEntityMagicDoorway) {
-            return ((TileEntityMagicDoorway) tileEntity).getBaseBlockState().getBlockHardness(worldIn, pos);
-        }
-        return super.getBlockHardness(blockState, worldIn, pos);
-    }
-
-    @Override
-    public int getHarvestLevel(IBlockState state) {
-        return -1;
-    }
-
-    @Nullable
-    @Override
-    public String getHarvestTool(IBlockState state) {
-        return null;
-    }
-
-    @Override
     public SoundType getSoundType(IBlockState state, World world, BlockPos pos, @Nullable Entity entity) {
         TileEntity tileEntity = world.getTileEntity(pos);
         if (tileEntity instanceof TileEntityMagicDoorway) {
@@ -139,11 +87,6 @@ public class BlockMagicDoorway extends Block {
             return replacedBlock.getBlock().getSoundType(replacedBlock, world, pos, null);
         }
         return super.getSoundType(state, world, pos, entity);
-    }
-
-    @Override
-    public boolean isReplaceable(IBlockAccess worldIn, BlockPos pos) {
-        return false;
     }
 
     @Override
@@ -198,20 +141,6 @@ public class BlockMagicDoorway extends Block {
                 return state.getValue(OPEN_EAST_WEST) ? BlockFaceShape.UNDEFINED : BlockFaceShape.SOLID;
         }
         return super.getBlockFaceShape(worldIn, state, pos, face);
-    }
-
-    public BlockRenderLayer getRenderLayer() {
-        return BlockRenderLayer.CUTOUT_MIPPED;
-    }
-
-    @Override
-    public boolean isOpaqueCube(IBlockState state) {
-        return false;
-    }
-
-    @Override
-    public boolean isFullCube(IBlockState state) {
-        return false;
     }
 
     @Override
@@ -280,11 +209,6 @@ public class BlockMagicDoorway extends Block {
         return super.getActualState(state, worldIn, pos);
     }
 
-    @Override
-    public boolean hasTileEntity(IBlockState state) {
-        return true;
-    }
-
     @Nullable
     @Override
     public TileEntity createTileEntity(World world, IBlockState state) {
@@ -314,38 +238,5 @@ public class BlockMagicDoorway extends Block {
         }
 
         return result;
-    }
-
-    /**
-     * The doorway has two parts: top and bottom.
-     */
-    public enum EnumPartType implements IStringSerializable {
-        TOP("top", 0),
-        BOTTOM("bottom", 1),
-        ;
-
-        private final String name;
-        private final int value;
-
-        /**
-         * @param name  The name of the part.
-         * @param value The integer value of the part; used for setting block metadata.
-         */
-        EnumPartType(String name, int value) {
-            this.name = name;
-            this.value = value;
-        }
-
-        @Override
-        public String getName() {
-            return name;
-        }
-
-        /**
-         * @return The integer value of the part; used for setting block metadata.
-         */
-        int getValue() {
-            return value;
-        }
     }
 }
